:sparkles: Adust Wallet Controller DTO

+++ src/main/java/recargapay/wallet/application/dto/response/DepositResponseDTO.java
@@ -1,4 +1,6 @@
 package recargapay.wallet.application.dto.response;
 
-public record DepositResponseDTO(String message, Double amount) {
+import java.math.BigDecimal;
+
+public record DepositResponseDTO(String message, BigDecimal amount) {
 }
